3차 목표 달성! - 각 파일에 컴포넌트 하나씩 담기
- 각 파일에 동일한 이름의 컴포넌트를 하나씩 담아 과제를 완성했습니다.
